[FIX] removed useless type conversion

+++ src/main/java/it/unibo/project/controller/core/impl/LauncherImpl.java
@@ -33,9 +33,9 @@ public final class LauncherImpl implements Launcher {
      */
     public static final Launcher LAUNCHER = new LauncherImpl();
     /** {@code width} in pixel of window. */
-    public static final int WIDTH = (int) GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice().getDisplayMode().getWidth() ;
+    public static final int WIDTH = GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice().getDisplayMode().getWidth() ;
     /** {@code height} in pixel of window. */
-    public static final int HEIGHT = (int) GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice().getDisplayMode().getHeight();
+    public static final int HEIGHT = GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice().getDisplayMode().getHeight();
     /** {@code quantity} of cell in {@code orizontal} direction. */
     public static final int ORIZ_CELL = 15;
     /** {@code quantity} of cell in {@code vertical} direction. */
